abstract out printErrors
